MOve disable command palette hook into ui
