+++ frontend/src/metabase/ui/hooks/use-disable-command-palette.tsx
